11th Commit
Initialized npm in the project and created a package.json file. Added 'start' and 'start-server' scripts to the scripts section of the package.json file.
These scripts are used to start the application by running 'node app.js'. The 'start' script can be run with 'npm start', and the 'start-server' script can be run with 'npm run start-server'. This allows for a more convenient way to start the application, especially when sharing the project with others.