118th Commit
feat: Add order functionality

Implemented the addOrder method in the User model. This method inserts the current user's cart into a new 'orders' collection in the database, representing the user's order. After the order is successfully inserted, the user's cart is cleared both in the User model and in the database.

The addOrder method is called in the postOrder function in the shop controller, replacing the previous order creation logic. After the order is successfully created, the user is redirected.

This change introduces a new way of creating relations in the application. Unlike the cart, which is an embedded document in the User model, orders are stored in a separate collection. This is because a user can have many orders, and embedding them all in the User model could make the user objects too large.

This commit includes:
- Creation of a new 'orders' collection in the database.
- Implementation of the addOrder method in the User model.
- Modification of the postOrder function in the shop controller to use the new addOrder method.

Note: The orders page route needs to be re-added to the application, as it was removed previously. This will be addressed in a future commit.