119th Commit
feat: Enhance order data

Enhanced the data stored in each order. Each order now includes not only the items in the cart at the time of the order, but also some information about the user who made the order. This includes the user's ID, name, and email. This data is stored as an embedded document within each order.

Additionally, the items in the order now include more information about each product, not just the product ID. This is achieved by calling the getCart method in the addOrder method, which fetches detailed information about each product in the cart.

This commit includes:
- Modification of the addOrder method in the User model to include more data in each order.
- Fetching of detailed product information in the addOrder method.

Note: The getOrders method still needs to be implemented. This will be addressed in a future commit.