123rd Commit
at: Implement Mongoose model for Product and update admin controller

Created a Mongoose model for the Product schema in the 'product.js' model file. This model is exported and used in the 'admin.js' controller.

Updated the 'postAddProduct' function in the 'admin.js' controller to create a new product using the Mongoose model. The product is created with a JavaScript object that maps the schema fields to the data received in the controller action.

Commented out the user-related code in 'app.js' as we currently do not have a user model.

Note: The use of Mongoose models allows us to create and manage instances of our data. The 'save' method provided by Mongoose is used to save the new product to the database.