124th Commit
feat: Update shop controller to fetch products using Mongoose

Updated the 'getProducts' and 'getIndex' functions in the 'shop.js' controller to fetch all products using the 'find' method provided by Mongoose. This method is called on the Product model, which is imported from the 'product.js' model file.

Note: The 'find' method in Mongoose returns an array of all matching documents. For large amounts of data, it is recommended to use a cursor or limit the set of data that is retrieved.