125th Commit
feat: Update shop controller to fetch a single product using Mongoose

Updated the 'getProduct' function in the 'shop.js' controller to fetch a single product using the 'findById' method provided by Mongoose. This method is called on the Product model, which is imported from the 'product.js' model file, with the product ID as an argument.

Uncommented the route for loading a single product in the 'shop.js' router file.

Note: The 'findById' method in Mongoose accepts a string as an argument and automatically converts it to an ObjectId. This provides convenience as we do not have to manually convert the string to an ObjectId.