126th Commit feat: Update admin controller to edit products using Mongoose
Updated the 'postEditProduct' function in the 'admin.js' controller to edit a product using Mongoose. The product is fetched using the 'findById' method and then updated by setting the fields to the new values. The 'save' method is then called on the product to save the changes to the database.

Uncommented the 'postEditProduct' route in the 'admin.js' router file.

Note: Calling 'save' on an existing Mongoose object updates the object in the database instead of creating a new one. This provides convenience as we do not have to manually handle the update operation.