127th Commit
feat: Update admin controller to delete products using Mongoose

Updated the 'postDeleteProduct' function in the 'admin.js' controller to delete a product using the 'findByIdAndRemove' method provided by Mongoose. This method is called on the Product model, which is imported from the 'product.js' model file, with the product ID as an argument.

Uncommented the 'postDeleteProduct' route in the 'admin.js' router file.

Note: The 'findByIdAndRemove' method in Mongoose deletes the document matching the provided ID from the database. This provides convenience as we do not have to manually handle the delete operation.