129th Commit
feat: Add user-product relationship and store user ID with products

Updated the Product schema to include a 'userId' field, which is a reference to the User model. This field is used to establish a relationship between a product and a user.

Updated the 'postAddProduct' function in the 'admin.js' file to store the user ID when a new product is created. The user ID is obtained from the request object and passed to the Product constructor.

Note: Mongoose allows storing the entire user object in the 'userId' field, and it will automatically pick the ID from that object. This feature is used in the 'postAddProduct' function.