149th Commit
Implement user sign-up flow.
This includes creating a sign-up form, a post sign-up route, and a post sign-up action.
The action validates user input and checks for existing users before creating a new user in the database.
The user model was also updated to include a password field.
Removed dummy user from the database.

Realised there's an issue in the newly created user, will be addressed and fixed in next commit