152nd Commit
Implement route protection

This commit introduces route protection to prevent unauthorized access to certain pages. Previously, even after logging out, users could manually enter URLs to access pages like 'admin add product'. This has been addressed by checking the user's authentication status before rendering pages.

In the admin controller, a check has been added in the 'get add product' function to verify if the user is logged in before rendering the page. If the user is not logged in, they are redirected to the login page. This prevents the 'add product' page from being loaded by users who are not logged in.

This is a preliminary implementation of route protection. A more scalable solution will be introduced in subsequent commits.